refactor: move script to a separate package
