+++ apps/public-docsite-v9/storybook-llms.config.js
@@ -1,0 +1,18 @@
+// @ts-check
+
+const storybookConfig = require('./.storybook/main');
+
+/**
+ * Config for the LLMs docs generator script.
+ * @see {@link file://./../../tools/storybook-llms-extractor/src/cli.ts}
+ *
+ * @type {import('@fluentui/storybook-llms-extractor').Args}
+ */
+module.exports = {
+  distPath: './dist/storybook',
+  baseUrl: 'https://react.fluentui.dev',
+  summaryTitle: 'Fluent UI React v9',
+  summaryDescription:
+    "Fluent UI React is a library of React components that implement Microsoft's [Fluent Design System](https://fluent2.microsoft.design).",
+  refs: storybookConfig.refs ? Object.values(storybookConfig.refs) : [],
+};
