+++ scripts/storybook/tsconfig.json
@@ -1,7 +1,6 @@
 {
   "extends": "@tsconfig/node18/tsconfig.json",
   "compilerOptions": {
-    "lib": ["ES2019", "DOM"],
     "target": "ES2019",
     "pretty": true,
     "noEmit": true,
