+++ tools/storybook-llms-extractor/bin/storybook-llms-extractor.js
@@ -1,0 +1,5 @@
+#!/usr/bin/env node
+
+// @ts-check
+
+require('../dist/src/cli');
