+++ tools/storybook-llms-extractor/project.json
@@ -1,0 +1,19 @@
+{
+  "name": "storybook-llms-extractor",
+  "$schema": "../../node_modules/nx/schemas/project-schema.json",
+  "sourceRoot": "tools/storybook-llms-extractor/src",
+  "projectType": "library",
+  "tags": ["platform:any", "tools"],
+  "targets": {
+    "build": {
+      "executor": "@nx/js:swc",
+      "outputs": ["{options.outputPath}"],
+      "options": {
+        "outputPath": "{projectRoot}/dist",
+        "main": "tools/storybook-llms-extractor/src/index.ts",
+        "tsConfig": "tools/storybook-llms-extractor/tsconfig.lib.json",
+        "assets": ["tools/storybook-llms-extractor/*.md", "tools/storybook-llms-extractor/src/template/*"]
+      }
+    }
+  }
+}
