+++ tools/storybook-llms-extractor/src/__fixtures__/args.ts
@@ -1,0 +1,15 @@
+import type { Args } from '../types';
+
+export const argsWithRefs = {
+  summaryTitle: 'Fluent UI React v9',
+  summaryDescription:
+    "Fluent UI React is a library of React components that implement Microsoft's Fluent Design System.",
+  baseUrl: 'https://react.fluentui.dev',
+  distPath: 'dist/storybook',
+  refs: [
+    {
+      title: 'Charts v9',
+      url: 'https://charts.fluentui.dev',
+    },
+  ],
+} satisfies Args;
