+++ tools/storybook-llms-extractor/src/__fixtures__/html/subheadings.html
@@ -1,0 +1,12 @@
+<h2 id="setup">
+  <a aria-hidden="true" href="#setup" tabindex="-1" target="_self"></a>
+  Setup
+</h2>
+<h3 id="react-18">
+  <a aria-hidden="true" href="#react-18" tabindex="-1" target="_self"></a>
+  React 18
+</h3>
+<h4 id="ssr-with-nextjs">
+  <a aria-hidden="true" href="#ssr-with-nextjs" tabindex="-1" target="_self"></a>
+  SSR with Next.js
+</h4>
