+++ tools/storybook-llms-extractor/src/index.ts
@@ -1,0 +1,3 @@
+export type { Args } from './types';
+
+export {};
